+++ packages/cli/src/commands/migrate/index.ts
@@ -1,9 +1,16 @@
 #!/usr/bin/env node
+import { resolve } from 'path';
 import { Command } from 'commander';
 import { Listr } from 'listr2';
+import { createLogger, format, transports } from 'winston';
 
 import { version } from '../../../package.json';
-import { parseCommaSeparatedList, gitIsRepoDirty, resetFiles } from '../../utils';
+import {
+  parseCommaSeparatedList,
+  gitIsRepoDirty,
+  resetFiles,
+  getLintConfigPath,
+} from '../../utils';
 import {
   initTask,
   depInstallTask,
@@ -11,6 +18,7 @@ import {
   tsConfigTask,
   lintConfigTask,
   createScriptsTask,
+  regenTask,
 } from './tasks';
 
 import type { MigrateCommandOptions } from '../../types';
@@ -37,17 +45,22 @@ migrateCommand
   .option('-i, --interactive', 'interactive mode')
   .option('-v, --verbose', 'print debugging logs')
   .option('-d, --dryRun', 'print files that will be attempted to migrate', false)
+  .option('-r, --regen', 'print out current migration status')
   .action(async (options: MigrateCommandOptions) => {
     await migrate(options);
   });
 
 async function migrate(options: MigrateCommandOptions): Promise<void> {
-  console.log(`@rehearsal/migrate ${version.trim()}`);
+  const loggerLevel = options.verbose ? 'debug' : 'info';
+  const logger = createLogger({
+    transports: [new transports.Console({ format: format.cli(), level: loggerLevel })],
+  });
+  logger.info(`@rehearsal/migrate ${version.trim()}`);
 
-  if (!options.dryRun) {
+  if (!options.dryRun && !options.regen) {
     const hasUncommittedFiles = await gitIsRepoDirty(options.basePath);
     if (hasUncommittedFiles) {
-      console.warn(
+      logger.warn(
         'You have uncommitted files in your repo. Please commit or stash them as Rehearsal will reset your uncommitted changes.'
       );
       process.exit(0);
@@ -72,15 +85,31 @@ async function migrate(options: MigrateCommandOptions): Promise<void> {
       // For issue #549, have to use simple renderer for the interactive edit flow
       // previous ctx is needed for the isolated convertTask
       const ctx = await new Listr(tasks, defaultListrOption).run();
-      await new Listr([await convertTask(options, ctx)], {
+      await new Listr([await convertTask(options, logger, ctx)], {
         renderer: 'simple',
         ...defaultListrOption,
       }).run();
+    } else if (options.regen) {
+      const lintConfigPath = getLintConfigPath(options.basePath);
+      if (!lintConfigPath) {
+        logger.warn(
+          `${lintConfigPath} does not exist. You need to run rehearsal migrate first before you can run rehearsal migrate --regen`
+        );
+      }
+      const tsConfigPath = resolve(options.basePath, 'tsconfig.json');
+      if (!tsConfigPath) {
+        logger.warn(
+          `${tsConfigPath} does not exist. You need to run rehearsal migrate first before you can run rehearsal migrate --regen`
+        );
+      }
+
+      const tasks = new Listr([await regenTask(options, logger)], defaultListrOption);
+      await tasks.run();
     } else {
-      await new Listr([...tasks, await convertTask(options)], defaultListrOption).run();
+      await new Listr([...tasks, await convertTask(options, logger)], defaultListrOption).run();
     }
   } catch (e) {
     await resetFiles();
-    console.error(`${e}`);
+    logger.error(`${e}`);
   }
 }