+++ packages/cli/src/commands/migrate/tasks/index.ts
@@ -4,3 +4,4 @@ export * from './dependency-install';
 export * from './initialize';
 export * from './config-ts';
 export * from './create-scripts';
+export * from './regen';