+++ packages/regen/test/fixtures/src/test1.ts
@@ -1,0 +1,1 @@
+export const DAY_IN_WEEK = 7